Comment out the postgres configuration from .env.template (#502)
<!-- .github/pull_request_template.md -->

## Description
Simplest cognee docker setup of SQLite-NetworkX-LanceDB should not
enable postgres configuration in the .env.template by default. I think
leaving postgres details commented is better than removing them
entirely. This keeps simple optionality visible for newcomers.

## DCO Affirmation
I affirm that all code in every commit of this pull request conforms to
the terms of the Topoteretes Developer Certificate of Origin



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Chores**
- Updated configuration template to remove database parameters not used
in the default setup, with clearer guidance to ensure the intended
configuration is maintained.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->