fix: entrypoint.sh to not fail on first docker up, improved handling of migrations, signals and errors. (#546)
<!-- .github/pull_request_template.md -->

## Description
<!-- Provide a clear description of the changes in this PR -->
In it's current form, the entrypoint.sh script will run but fail with
exit code 3 on the first docker compose up. Technically, running docker
compose up a second time will not throw the same error and the
application works fine. The new changes will improve the first time user
experience and improve on some other aspects.

Summary of Changes:
1- entrypoint.sh to not fail with exit code 3 on first docker up.
2- Improved error and signal handling with set -e.
3- Improved database migration, verification and error handling. Avoids
schema version mismatch and ensures db schema is always in sync with
application code.
4- Added exec before Gunicorn commands to ensure proper signal handling.

## DCO Affirmation
I affirm that all code in every commit of this pull request conforms to
the terms of the Topoteretes Developer Certificate of Origin


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Chores**
	- Improved error handling for smoother database migrations and startup.
	- Updated process management to ensure reliable application launch.
- Optimized worker configuration and introduced a startup delay to
guarantee database readiness.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: soekja <soekja@users.noreply.github.com>
Co-authored-by: soekja <soekja@users.noreply.github.com>
Co-authored-by: Vasilije <8619304+Vasilije1990@users.noreply.github.com>